🎨 Improved member filter option queries and extracted MembersListItem
Increased labels, tiers, and newsletter filter queries from default page
size (~15) to 100 to prevent silently missing options on sites with many
labels. Extracted member list item rendering into a dedicated component.
